refactor: controller 계층 분리
Application class에서 controller 계층을 분리
